requirements.txt
django==3.0.5
Pillow==7.1.2
django-crispy-froms==1.9.0
djangorestframework==3.11.0
model-mommy==2.0.0
coverage==5.1

#con el comando -pip list- podemos ver todas las versiones de los modulos instalados